rename static images folder to avoid conflict
