+++ frontend/app/[city]/massnahmen/[...task]/page.tsx
@@ -6,7 +6,7 @@ import { Container } from "react-bootstrap";
 import styles from "../page.module.scss";
 import Image from "next/image";
 import { Task, useGetTasksByCity } from "@/app/TasksService";
-import arrow from "../../../../public/images/arrow-right-down.svg";
+import arrow from "../../../../public/imgs/arrow-right-down.svg";
 import Breadcrumb from "@/app/components/BreadCrumb";
 
 const getTaskBySlug = (tasks: Task[] | undefined, taskSlug: string): Task | undefined => {
