+++ frontend/app/components/ExecutionStatusIcon.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import Image from "next/image";
 
-import abgeschlossen from "../../public/images/icon-abgeschlossen.svg";
-import gescheitert from "../../public/images/icon-gescheitert.svg";
-import inArbeit from "../../public/images/icon-in_arbeit.svg";
-import unbekannt from "../../public/images/icon-unbekannt.svg";
-import verzoegert from "../../public/images/icon-verzoegert_fehlt.svg";
+import abgeschlossen from "../../public/imgs/icon-abgeschlossen.svg";
+import gescheitert from "../../public/imgs/icon-gescheitert.svg";
+import inArbeit from "../../public/imgs/icon-in_arbeit.svg";
+import unbekannt from "../../public/imgs/icon-unbekannt.svg";
+import verzoegert from "../../public/imgs/icon-verzoegert_fehlt.svg";
 
 enum ExecutionStatus {
   UNKNOWN = 0,
